Remove jersey-json and jackson dependencies.
These libraries are not used anymore.
JSON serialization and deserialization is meanwhile performed by our custom provider classes.
(Jackson in turn was a dependency of jersey-json)

Refactor Files plugin in the way it no longer depends on jersey-json.
(It was the last and only plugin that relied on jersey-json).

For plugins that provide a REST service added jersey-core dependency to their POMs.
These plugins rely on the javax.ws.rs API which is included in jersey-core.
FIXME: we could include the sole JAX-RS API as a compile time dependency instead of relying on a specific implementation (Jersey).