Fix deepamehta3-plugin-example tests
Added OSGi log service to Felix provision POM. The log service is required by Pax Exam to run the deepamehta3-plugin-example tests.

Is the Felix provision POM the right place to add the log service? The log service is *only* required by Pax Exam. It is *not* (yet) required for the regular runtime environment (development and distribution).