Introduce new Core concept: Directives
Directives are processing instructions for clients.
Directives are constructed at server-side. They are returned by core service calls.
Plugins take part in the directive construction process.

Through directives the core serive calls delivers a more expressive response that is useful for clients to update their model and GUI. When the client sends e.g. an "update association" request, the service might respond not only with the updated association but additionally with a "update topic type" directive to instruct the client to update its type cache. This would be the case if the changed association affects the semantic of a type definition.

Advantage for the Type Editor plugin: when an association that is involved in a type definition is created or retyped the client's type definition is immediately up-to-date. No manual "update" operation is required anymore. Example: when associating the "Book" type with a "ISBN" type, and typing the association to "Composition" or "Aggregation", the ISBN field is immediately available in book pages and forms.

The directives semantic is client-independant. Every type of client can benefit from them.

The directives concept is similar to the "Command" pattern (although the implementation is very simple for the moment). In fact the Directive class could be named Command but this term is already in use.

Perspective:
a) API: the core service calls return values might be gradually transformed to "Directives".
b) The Directives concept might render the "Topic Enrichment" and "Type Enrichment" concept obsolete.
c) Directives might be beneficial for a future undo feature