Webbrowser: visualize links on canvas (experiment)
This feature is called "Trailblazing"

For this to work you must ensure that the (URL topic's) URL ends with a / in case it points to a directory, e.g. www.deepamehta.de/

Furthermore:
Adapt File plugin's proxy service to DM4.
Accessible via /proxy (formerly /resource).

Note: The proxy service is required for the "Trailblazing" feature to work because of the browser's security restrictions.