Webclient: fix scrolling menus (#71).
When triggering a stateful menu the current selection is a) placed under the mouse pointer (as before) and b) is immediately highlighted. No extra mouse move is required for highlighting.

One issue remains: placing the selection under the mouse pointer doesn't work if the menu triggering click is performed on the menu's little triangle icon.

See ticket 71.