Files module: basically functional (#78)
Functionality like in DM3: Drop a bunch of files and folders from your filebrowser to the DM canvas. They are now represented as topics (of type "File" resp. "Folder").

There are 2 features:

1) to open the file in its original application doubleclick the file topic.

2) to display the file contents right in DM's content panel click the file topic.
This is currently supported for:
- Plain text
- Images
- Audio
- Video
- PDF

It should work with Firefox and Safari (Chrome not yet supported).
Audio/Video/PDF capabilities rely on your browser plugins. Certainly there is e.g. a Firefox PDF plugin for your OS available on the net.

IMPORTANT for Firefox: for the DM file handling capability to work you must set one particular preference. Type about:config in the Firefox address bar, then double-click the signed.applets.codebase_principal_support preference to set it to true (no Firefox restart required). Later on, when you drop a file to the canvas Firefox will ask you for permission to read the file. Click "Allow".
(If you not set the particular preference, or if you click "Deny", Firefox has no access to your local files for security reasons.)
This is only a workaround. Later versions of DM will not require this setting.

With Safari no extra setup is required.

See ticket 78.