Core service: crucial API change of createTopic().
The core service's createTopic() call returns a set of directives instead of the created topic directly.

This allows plugins to send directives back to the client as reaction on a created (child) topic. This works also if the create operation is a consequence of a complex update request.

Example: the user enters an address in a Person form (Address is a child type of Person). This results in an update request. As a consequence an Address topic is created at server-side (provided the Person had no Address assigned before). The Geomaps module reacts on the Address creation: it looks up the geo coordinate (via geocoder), attaches a corresponding facet to the address, and sends an ADD_MARKER directive back to the client. The client displays a marker on the geomap.

The changed createTopic() signature also affects the REST API. If you create topics via REST API (by POSTing to /core/topic) you must do some extra parsing on the response to find the topic created: it is the argument of the *last* CREATE_TOPIC directive in the returned set of directives. See the de.deepamehta.core.service.Directives class for the directives serialization format. Furthermore its getCreatedTopic() method illustrates how to find the topic created.

Another core extension: modules can define custom directives.

They do so by definig an enum that implements the Directive interface. Directive is now an interface. It contains just one name() method. See an example in GeomapsPlugin.java:

    public enum GeomapDirective implements Directive {
        ADD_MARKER
    }

For processing custom directives the Webclient provides a "process_directive" hook. See an example in geomaps_plugin.js:

    this.process_directive = function(directive) {
        if (directive.type == "ADD_MARKER") {
            ...
        }
    }