Webclient: access split panel's slider (#148).
Plugins can set/get the split panel's slider position programmatically.

Set the slider's position:
{{{
    dm4c.split_panel.set_slider_position(800)
}}}
The argument is the width of the left panel.

To set the slider to the rightmost position you can do something like this:
{{{
    dm4c.split_panel.set_slider_position(window.innerWidth - 36)
}}}

However, it's not optimal: unwanted window scroll appear as the content panel's Edit button still occupies space. Even if the content panel is empty (no topic selected), there are still some margins which occupy space.

So, setting the slider to the rightmost position is probably not the proper approach for your case, which actually is: show/hide the content panel. There is currently no support for doing this (as far as I know). You could open a separate ticket.

Get the slider's position:
{{{
    var pos = dm4c.split_panel.get_slider_position()
}}}

See ticket 148.