Core: model-driven composite value updating.
An topic/association update request can contain external (non-model-compliant) composite values. These are ignored by the core and passed to the plugins for processing.

This allows plugins to extend an update request by proprietary data and to provide the corresponding logic to process that data.

This feature is exploited e.g. in conjunction with the Facets plugin.

IMPORTANT for developers and REST API users: the syntax of update requests that contain topic *references* (as used in conjunction with aggregations) has changed. The "$id" qualifier is renamed to "ref_id:" and moved from the key-part to the value-part. Example:

Old syntax:

    {"dm4c.contacts.city$id": 2365}

New syntax:

    {"dm4c.contacts.city": "ref_id:2365"}