Core: change object API.
In interface de.deepamehta.core.DeepaMehtaObject: replace method
    update(DeepaMehtaObjectModel, ClientState, Directives)
by
    updateCompositeValue(AssociationDefinition, TopicModel, ClientState, Directives)

Anyway, both methods are low-level and should not to be part of the public API. They are not supposed to be called by plugin developers and will be vanish in the future. For the moment they are required by the external Facets module. This suggests the Facets module for being incorporated into the DeepaMehta core.