Core fix: catch invalid association defs (#253).
The webclient doesn't break anymore under the condition decribed in #253.
However, the fix is not complete yet: once the user has repaired the assoc def it is not active before she performs one extra step: retype the association, and then retype it back to the desired association type.

See ticket 253.