Core: by-URI topic refs in update requests (#267)
... as well as in create requests.

REST interface: to refer to an existing topic besides the ref_id: prefix you can now use the ref_uri: prefix as value in a composite structure.

DM Core Java API: when programmatically constructing a TopicModel to be used in a create or update call use the CompositeValue's new methods:
    put_ref(String childTypeUri, long refTopicId)
    put_ref(String childTypeUri, String refTopicUri)

Close ticket 267.