Core: dynamic un/register Jersey servlet (#272).
Once all REST resources are removed the Jersey servlet is unregistered.
As soon as resources are added the Jersey servlet is registered again.
So, actions like reloading the DM core or shutdown the platform perform properly.

Note: there is still an issue with reloading the Webclient module but this is not REST related and needs to be addressed in another ticket.

See ticket 272.