Webclient: plugin namespace convention (#275).
IMPORTANT for plugin developers:
For proper namespacing there is the convention to use the plugin's URI when registering via dm4c.add_plugin():

    dm4c.add_plugin("de.deepamehta.topicmaps", function() {
        ...
    })

Technically this string is arbitrary. It's just the identifier by which you can access the plugin later on, usually to call its public API methods:

    dm4c.get_plugin("de.deepamehta.topicmaps").get_topicmap()

However, the string must be globally unique, so it is most natural to use the plugin's URI. Actually the plugin URI is the plugin bundle's symbolic name, which is calculated from the plugin project's group ID and artifact ID.

See ticket 275.

Furthermore in Webclient: exceptions that occur while loading and executing a script via dm4c.load_script() are rethrown to let them appear in the browser's error console.