Webclient: the Checkbox Renderer works (#270).
The (form rendering aspect of the) Checkbox Renderer allows the user to select one or more items. The set of possible items comprises all the topic instances of a certain type -- the item type. The item type must be aggregated to a parent type. If a parent's type instance is rendered the checkboxes appear.

To use the Checkbox Renderer enter "dm4.webclient.checkbox_renderer" in the item type's "Multi Renderer URI" view configuration.

Note: for the Checkbox Renderer to work 2 requirements must be met:
    1) the item type is associated to its parent type via "Aggregation Definition"
       It's in the nature of the Checkbox Renderer that "Composition Definition" doesn't work.
    2) parent type -> item type must be a "Many" relationship.
       Again, it's in the nature of the Checkbox Renderer that "One" makes no sense.

Technically the Checkbox Renderer is a "Multi" renderer (new concept).
A multi renderer is responsible for rendering the multiple values resulting from a "Many" relationship.

Layout and other minor issues remain.

See ticket 270.