Webclient: revise renderer file structure.
Plugins which provide renderers adhere to a new file structure:

script/
    renderers/
        page_renderers/
        simple_renderers/
        multi_renderers/
        canvas_renderers/

2 things are new here:
    - the "renderers" directory
    - "field_renderers" is now named "simple_renderers"

The term "field renderer" is now obsolete. We're talking of "simple renderers" now.
("composite renderes" are about to come, and thus the new wording is conform with "simple" and "composite" topics.)

The new wording is reflected at several spots:
    - view config setting:    "Field Renderer URI"                ->  "Simple Renderer URI"
    - the corresponding URI:  "dm4.webclient.field_renderer_uri"  ->  "dm4.webclient.simple_renderer_uri"
    - loading method:         dm4c.load_field_renderer()          ->  dm4c.load_simple_renderer()
    - registration call:      dm4c.add_field_renderer()           ->  dm4c.add_simple_renderer()

IMPORTANT: plugin developers must adapt to the new file structure and the new wording.