Webclient/Topicmaps: change renderer wording.
The concept of "Canvas Renderer" is now called "Topicmap Renderer".
Because the respective renderer API is actually topicmap-centered (add_topic(), add_association(), ...) Topicmap Renderer is the more appropriate name.

The Webclient's interface is renamed:
    CanvasRenderer -> TopicmapRenderer

The default implementation is renamed:
    TopicmapRenderer" -> "DefaultTopicmapRenderer

The renderer directory is renamed:
    "canvas_renderers" -> "topicmap_renderers"

The client-side event (hook) is renamed:
    "canvas_renderer" -> "topicmap_renderer"

The (hidden in GUI) model type is renamed, along with its URI:
    "Canvas Renderer URI" -> "Topicmap Renderer URI"
    "dm4.topicmaps.canvas_renderer_uri" -> "dm4.topicmaps.topicmap_renderer_uri"

IMPORTANT: because of the model change you have to reset the DB one more time.
IMPORTANT: plugins which provide topicmap renderers must be adapted to the name changes mentioned above.