Begin of new feature: File Upload (#289).
The "Create" menu provides a new "Upload File" item. It opens a file upload dialog. The choosen file is uploaded to the server. It is not yet processed.

Plugin developers can invoke the file upload dialog programmatically:
    dm4c.upload_dialog.show(command, callback)
(Both parameters are not yet used.)

The REST API of the "Files" plugin provides a new method:
    void uploadFile(UploadedFile file, String command, ClientState clientState)

A multipart/form-data request with the file content as its body can be POSTed to /files/{command}
(The command is not yet used.)

1 class is moved from core "util" to "service" package:
    UploadedFile

The entire "Commands" concept is finally dropped from core.
2 classes are removed from the "service" package:
    CommandParams
    ComandResult
2 provider classes are removed from the "Webservice" plugin:
    CommandParamsProvider
    CommandResultProvider

See ticket 289