Files plugin: store uploaded files (#289).
Try it via "Upload File" from the Create menu.
The uploaded file is stored in "deepamehta-files" (next to "deepamehta-db").
The server response is shown in an alert.

There is a new property in global POM:

    dm4.files.path

This is the path of the server-side file repository.
By default it is named "deepamehta-files" and stored next to "deepamehta-db".

The response of the upload request is an JSON object.
For the moment it contains just the name of the uploaded file.

    {
        filename: ...
    }

This JSON object is passed to the callback function that is passed to the show() method:

    dm4c.upload_dialog.show(callback)

The 2nd show() parameter ("command") is dropped.
The URI for POSTing a file is just /files (the "command" path parameter is dropped).

For the upcoming 4.0.12 release the file repository is not realized in a separate module (#291) but
an integral feature of the "Files" plugin. However, it is encapsulated in a new class: FileRepository

The class UploadedFile has moved from Core module to Files plugin.