Webclient: revise GUIToolkit's Dialog API.
The dm4c.ui.dialog() call returns a Dialog object. This object provides these methods:
{{{
    open()
    close()
    empty()
    appen(element)
}}}
Thus, the plugin developer is provided with a high-level object-orientated API and must no longer (low-level) operate on the DOM.

Furthermore the dm4c.ui.dialog() call takes a single "config" parameter. So, the developer is no longer required to pass all the 6 parameters (in the proper order!). Instead the config object can have up to 6 properties, whereas all are optional. In particular the developer is no longer required to pass a unique ID for every dialog:
{{{
    id             - Optional: the dialog content div will get this ID. Useful to style the
                     dialog with CSS. If not specified no "id" attribute is set.
    title          - Optional: the dialog title. If not specified an empty string is used.
    content        - Optional: the dialog content (HTML or jQuery objects). If not specified
                     the dialog will be empty. It still can have a title and/or a button.
    width          - Optional: the dialog width (CSS value). If not specified "auto" is used.
    button_label   - Optional: the button label. If not specified no button appears.
                     Note: the button label and button handler must be set together.
    button_handler - Optional: the button handler function. If not specified no button appears.
                     Note: the button label and button handler must be set together.
}}}

IMPORTANT: if your plugin makes use of dm4c.ui.dialog() it must be adapted.

Another hint for plugin developers:
A plugin (with a client-side component) is not strictly required to have a main file (plugin.js)
This is perfect for plugins whose only purpose is to provide custom renderers.
This applies e.g. to the Icon Picker plugin.