Core: publishing a directory to the web (#292).
The Core's WebPublishingService provides a new feature for publishing a directory of the server's file system to the web.
A plugin can do so by calling:
{{{
    publishDirectory(String directoryPath, String uriNamespace)
}}}
(This method is located in the plugin base class, PluginActivator).
Unpublishing is automatic when the plugin shuts down.

This feature replaces that very feature of the Proxy plugin.
The remaining Proxy features are about to be transfered to the Files plugin and the Proxy plugin is about to be dropped.

The first plugin who uses this new feature is the Files plugin. It publishes its file repository to URI namespace /filerepo
The file repository is still be configured via the POM properties
{{{
    dm4.proxy.files.path
    dm4.proxy.net.filter
}}}
These are about to be renamed.

CAUTION: the Proxy plugin's security netfilter feature is not yet ported to the WebPublishingService, so for the moment the "file-repo" branch opens a huge security hole. Do NOT install it!

See ticket 292.