Transfer Proxy features to Files plugin (#292).
The Proxy plugin and the /proxy URI namespace are obsolete.
The Proxy plugin is removed.
Its functionality is taken over by the Files plugin.
URI namespaces for the files API and the actual file repository are respectively:
{{{
    files/
    filerepo/
}}}

The 2 dm4.proxy... config properties are obsolete.
The new properties are:
{{{
    dm4.filerepo.path
    dm4.filerepo.netfilter
}}}

CAUTION: access to the file repository is not yet secured by the netfilter setting.
For single-user installations do NOT install the "file-repo" branch.
For multi-user installations make shure dm4.filerepo.path is set to a directory whose contents are supposed to be public.

See ticket 292.