Files plugin: consolidate security (#292, #289).
Both, the Files API (/files URI) and the file respository (/filerepo URI) are consistently secured.
Each of the respective requests is secured by 2 measures:
    - Subnet restriction: the request must originate from an allowed subnet (according to the dm4.filerepo.netfilter
      setting).
      Otherwise "403 Forbidden" is send back.
    - Fighting directory traversal attacks: once canonicalized (resolving ../ references) the request's "path"
      parameter must point inside the file repository (according to the dm4.filerepo.path setting).
      Otherwise "403 Forbidden" is send back.
Furthermore, if the request's "path" parameter refers to a non-existing resource (file or directory)
"404 Not Found" is send back.

Additionally the upload mechanism is revised: the ID of the respective File topic is not included in the name of the stored file anymore. The original (client-side) file name is only changed if such a file already exists in the file repository. In this case a -2, -3, -4, ... suffix is added to the file's basename.

See ticket 292.
See ticket 289.