Files plugin: create folder in file repo (#292).
GUI:
To create a folder in the file repository a Folder topic provides a "Create Folder" button.
The new folder is created inside the directory represented by that origin Folder topic.

API:
The Files OSGi service provides a new method:
{{{
    void createFolder(String folderName, String path)
}}}
"path" is relative to the file repo root dir.
It points to the directory where to create the new folder.
It must begin with "/".

Note:
for the moment creating repo folders is different from creating repo files (through upload) in 3 regards:
1) Creating a folder doesn't involve searching for a "free" name.
   If such a folder already exists 500 Internal Server Error is returned.
2) Creating a folder doesn't automatically create a corresponding Folder topic.
   The Folder topic is created on-the-fly when browsing the repo interactively.
3) Creating a folder returns no response entity, just 204 No Content.

See ticket 292.