Access Control: new role USER (#262).
There is a 5th role: a USER represents everyone who is logged in.
The USER role can be used in ACL entries just like the other roles.

This allows development of closed user group applications without the complexity of the MEMBER role and without the requirement of the Workspaces module to be installed.

See ticket 262.