Access Control: logout shuts down client (#293).
When logging out from a "private" DM installation the webclient shuts down, that is the browser window is emptied. This reflects the meaning of a "private" installation: the webclient is only accessible when logged in.

A DM installation is made "private" by setting the config property dm4.security.read_requires_login to true (in global pom.xml):
{{{
    <dm4.security.read_requires_login>true</dm4.security.read_requires_login>
}}}

Note this drawback: the next request logs in the user automatically. This is a known problem with HTTP Basic authentication.

See ticket 293.