Contacts: rename type "Email" to "Email Address".
Note: the type's URI is already "dm4.contacts.email_address". So, the renaming is purely cosmetic.

The renaming is intended to avoid confusion with the "Mail" type (part of dgf's external "DeepaMehta 4 Mail" plugin).

IMPORTANT: in an production environment a migration should be applied
or the "Email" type should be renamed interactively.

Furtermore: change MehtaGraph dependency from 1.0-SNAPSHOT to 1.0