Extend Access Control API (#262).
AccessControlService:
    - void setCreator(DeepaMehtaObject object, long usernameId) is a new method.
    - Topic getCreator(DeepaMehtaObject object)                 is a new method.
    - Topic getOwner(DeepaMehtaObject object)                   is a new method.
    - Topic getUsername(String username)                        is a new method.
    - DEFAULT_USERNAME is a public constant ("admin").
    - getOwnedTopic() is removed.

Permissions:
    - add() returns the Permissions object

The Access Control startup activity has changed: for topics and associations created by migrations NO creator and
NO access control entries are set. Because of the undeterministic plugin startup order it would be undeterministic
which topics and associations get these assignments and which one do not.

Instead it's up the plugins which create topic and asssociation instances in migrations to setup the creator and
access control entries. (this is now identical to the Workspace assignment behavoir.)

See ticket 262.