Core: revise aggregation update logic (#319).
Unchanged aggregations are not unnecessarily deleted (and restored) anymore.

Functional in principle, but not yet complete:
    - For many-aggregations the property panel is only up-to-date after topic reselection.
    - The checkbox multi-renderer is not yet adapted.

IMPORTANT for plugin developers:
To explicitly remove child topic assignments in many-aggregations you must include the del_id: prefix in the update request. On the other hand you're no longer required to include all the unchanged assignments in the request.

The del_id: prefix works now for both, compositions and aggregations.

For one-aggregations a new child topic assignment automatically removes an existing one. The del_id: prefix is not required here (and it is even not yet suported, that is, for one-aggregations you cant remove the existing single assignment.)

See ticket 319.