Core: idempotent child update operations (#319).
These update child operations are idempotent now:
    - delete child (composition many)
    - delete child assignment (aggregation many)
    - create child assignment (aggregation many)
So, an update request can contain e.g. del_id: prefixes to refer to child topics which are deleted already (compositions) resp. not assigned anymore (aggregations). This is not an error anymore. Instead, these partial update operations are ignored. The same applies to the ref_id: prefix when refering to a topic which is already assigned (aggregations).

Furthermore in Webclient:
The Checkbox Renderer is adapted to the Core's revised update logic.

CompositeValue:
    - put(String key, List<TopicModel> value) is dropped.
      Use add multiple values use add() consecutively.

Close ticket 319.