Webclient: shared renderer code (#335).
The code of the default topic renderer is refactored in order to be used by both, the topic renderer and the association renderer.

The plugin's renderer directories may contain subdirectories with supplemental code.
Loading the supplemental code is up to the plugin.

IMPORTANT for plugin developers:
If you call methods of the default topic renderer you must adapt your code.

Instead of
    var topic_renderer = dm4c.get_page_renderer("dm4.webclient.topic_renderer")
    topic_renderer.create_page_model(...)
    topic_renderer.render_page_model(...)
    topic_renderer.build_topic_model(...)
    topic_renderer.mode.INFO
    topic_renderer.mode.FORM
write
    dm4c.render.page_model.create_page_model(...)
    dm4c.render.page_model.render_page_model(...)
    dm4c.render.page_model.build_topic_model(...)
    dm4c.render.page_model.mode.INFO
    dm4c.render.page_model.mode.FORM

In preparation of "Association instance editor".

See ticket 335.