Webclient: display association info (#335).
An association's simple and composite values are displayed in the detail panel.
The corresponding form is pending.

New in Webclient API:
The Topic's and Association's "composite" property contains true Topic objects (formerly plain JavaScript objects).
This means the methods of the Topic prototype, e.g. get_type(), are available on objects accessed from a composite.
Calling the Topic constructor is not necessary anymore.
Example:
    type = topic.composite["my.uri"].get_type()

See ticket 335.