Core: refactor type construction pt.1 (#340).
Former approach:
Types are constructed with a model that is initialized only partial and then load and init the remaining parts themselfs.

New concept:
Construction of types is the responsibility of the object factory.
Loading the model and constructing the actual type object from it are strictly separated.
Never construct a type with a partly initialized model.

Does not yet work.
Do not install the "child-order" branch.

Done
    - data type
    - index modes
Pending
    - association definitions
    - label configuration
    - view configuration

In preparation of "Child Topic Order".

See ticket 340.