Webclient: reveal topic w/o selecting it (#387).
A styleguide is thoroughly applied to the webclient. It deals with how topics are revealed on the topicmap when clicked in the page panel:
    - when the *icon* is clicked the topic is revealed
    - when the *label* is clicked the topic is revealed AND selected
This provides for revealing a whole bunch of topics without loosing focus.

Webclient API:
    - dm4c.do_reveal_related_topic() has an optional 2nd parameter: "action".
      Its meaning is just the same as with dm4c.show_topic(): the action to perform on the revealed topic.
      3 possible values:
        "none" - do not select the topic (page panel doesn't change) -- the default.
        "show" - select the topic and show its info in the page panel.
        "edit" - select the topic and show its form in the page panel.
      If not specified (that is any falsish value) "none" is assumed.

INCOMPATIBILITY

When your plugin calls dm4c.do_reveal_related_topic() it must be adapted.
Just pass "show" as the 2nd parameter as "none" is the default now.
{{{
    dm4c.do_reveal_related_topic(topic_id, "show")
}}}

Furthermore if you pass a custom click handler to dm4c.render.topic_list() you should care about fulfilling the styleguide mentioned above. For that purpose a 2nd parameter is passed to your click handler: "spot". That is the spot where the topic has been clicked: "icon" or "label" (a string).

Example:
{{{
dm4c.render.topic_list(topics, function(topic, spot) {
    var action = spot == "label" && "show"
    dm4c.show_topic(topic, action)
})
}}}

See ticket 387.