More informative topic lists (#308, #329, #387).
Topic lists displayed in the page panel, e.g. search results, are more informative: topics which are already visible on the canvas are displayed with an disabled icon. This provides for better orientation when revealing a whole bunch of topics and is particularily useful in conjunction with #387.

See ticket 308.
See ticket 329.
See ticket 387.