Core: providing storage as OSGi service (#391).
Instead of instantiating through the factory the Core consumes the storage implementation as an OSGi service. A storage implmentation is expected to register DeepaMehtaStorage as an OSGi service.
The factory remain to exist in order to run the tests. While the test phase the OSGi environment is not running.

2 recently introduced config properties are dropped:
    - dm4.database.bundle
    - dm4.database.factory

Note: the storage can *not* be hot-deployed. Uninstalling the storage will not stop the Core. This is because the Core has technically no package dependency on the storage. The Core just depends on the storage service (DeepaMehtaStorage) which is defined in the Core package itself.
While being still a bundle the storage implementation is not a DeepaMehta plugin. It is not derived from PluginActivator (but from BundleActivator). The storage implementation provides a basis service required by the Core to run. Once the Core is up and running it acts as container for the various DM plugins.

See ticket 391.