Core fix: indexing of HTML fields (#391).
For HTML topics: the HTML tags are stripped before indexing.

BREAKING CHANGES
(for Storage implementers only)

DeepaMehtaStorage API:
    - storeTopicValue()       has additional "indexValue" parameter (optional)
    - storeAssociationValue() has additional "indexValue" parameter (optional)

So, there are separate 2 values provided: the value to be *stored* and the value to be *indexed*.
The latter one is optional. If not specified (null) the former one is used for both operations.

See ticket 391.